move frost.server.ext to frost.ext & remove unneeded headers on the client
