+++ frost/ext/__init__.py
@@ -1,0 +1,5 @@
+from frost.ext.cog import Cog
+from frost.ext.handler import Handler
+
+
+__all__ = ('Cog', 'Handler')
