+++ frost/ext/exceptions.py
@@ -1,6 +1,6 @@
 # Exceptions
 
 
-class DirectCogInstance(Exception):
+class DirectCogInstanceError(Exception):
     """Raised if the Cog class is directly instantiated.
     """
