--- frost/server/ext/__init__.py
@@ -1,5 +1,0 @@
-from frost.server.ext.cog import Cog
-from frost.server.ext.handler import Handler
-
-
-__all__ = ('Cog', 'Handler')
